added feedback
the player now receives feedback in the popup on as to why the chosen answer was incorrect. popup now has a text field to display this in. the json now takes one more argument being "explanation" make sure every statement has one (explanations in the json will be added next commit). the switch is now on at the beginning and has to be turned off by the player instead of on, this fits better with the text around it. the player can now also play single player without having to go through the setup of a multiplayer lobby. also fixed win and lose screen not working.